TINKERPOP-2639 Remove GryoLiteMessageSerializerV1d0
GryoLiteMessageSerializerV1d0 was deprecated in 3.2.6 almost as quickly as it was added as we came up with a better way to deal with detachment rather than pushing it down to the serializer. It is highly unlikely that this serializer is still being used especially since Gryo itself has been deprecated for GraphBinary.